+++ Basil/include/Platform/OpenGL/OpenGLShader.h
@@ -22,6 +22,10 @@ namespace Basil
 			void bind() const override;
 			void unbind() const override;
 			const std::string& getName() const override;
+			void setInt(const std::string& name, int value) override;
+			void setFloat3(const std::string& name, const glm::vec3& value) override;
+			void setFloat4(const std::string& name, const glm::vec4& value) override;
+			void setMat4(const std::string& name, const glm::mat4& matrix) override;
 			void uploadUniformInt(const std::string& name, int value);
 			void uploadUniformFloat(const std::string& name, float value);
 			void uploadUniformFloat2(const std::string& name, const glm::vec2& value);