+++ Sandbox/src/Sandbox2D.cpp
@@ -1,9 +1,8 @@
-#include "Platform/OpenGL/OpenGLShader.h"
 #include "Sandbox2D.h"
 
 #include <glm/gtc/matrix_transform.hpp>
 #include <glm/gtc/type_ptr.hpp>
-#include "imgui.h"
+#include <imgui.h>
 
 Sandbox2D::Sandbox2D() : Layer("Sandbox2D"), cameraController(1280.0f / 720.0f) {}
 
@@ -21,7 +20,7 @@ void Sandbox2D::onUpdate(Basil::Timestep timeStep)
 	Basil::Renderer::clear();
 
 	Basil::Renderer2D::beginScene(cameraController.getCamera());
-	Basil::Renderer2D::drawQuad({ 0.0f, 0.0f }, { 1.0f, 1.0f }, { 0.8f, 0.2f, 0.3f, 1.0f });
+	Basil::Renderer2D::drawQuad({ 0.0f, 0.0f }, { 1.0f, 1.0f }, squareColor);
 	Basil::Renderer::endScene();
 }